Final timing settings for the new LCD screen.
Front porch setting (both horizontal and vertical), have minimum
and maximum values according to the datasheet, and should be set
to the _typical_ setting in the datasheet. Once that is done, the
back porch settings can be used to to move the image horizontally
or vertically. The hsync-len, or pulse width setting, can be used
to remove 'ghost' images, and should be set in the range recommended
by the datasheet.