Add SPI-based SDCard support.
The GPIO_ACTIVE_HIGH is (counter-intuitively, actually specified by
value 0 and not value 1 for IMX processors). Added pullup and keeper
to set power pin in initial 'on' state. Keeper was added so that the
value is passed over from U-Boot to the actual kernel (U-Boot also
uses the device tree initially.)