Fix bug on serial port (ttymxc2).
Data went out, but incoming data caused a kernal panic. This was
due to changes in the driver, for the new 5.4 kernel. The waterlevel
mark is used elsewhere in a burst calculation, and it takes the value
and subtracts one, to get an 'aging' value. This value, however, cannot
be zero. We there adjust the value to be 2, as opposed to 1, which it was
patched to, under the previous kernel version.