Add RevC, RevD and RevE specific DTB files.
Also fix up 'label not found' error when building various device trees.

Note the UART changes for RevE:
-------------------------------
Native (high speed) UARTs swapped with expanded UARTs on DUART, so
that the SP1 and SP2 have the use of the more performant UARTs. This
is so that the EngineX board can operate at a higher frequency than
what is possible with the RevC.

Also ensure that SC16752 spi is enabled for RevC and RevE,
by setting the status for the SC16752 to "okay".

Also note the following fix for RevC and RevE:
----------------------------------------------
Move SCIS16752 clock definition to prevent clash with serdev.
With the original clock definition, the device was incorreclty
identified as a serdev device, and would then not correctly enumerate
as a tty. Moving the clock definition to not be a child node of the
device itself, fixes this issue.