Reduce trigger level for receive fifo to a single byte.
This improves the latency of the UART roundtrip communications,
as experienced with the engine board FireFly application. Initial
tests indicate around 60ms improvement.

NOTE: Test this, as the driver has changed from the previous
version and this may result in a latency regression.

NOTE: Rducing this value to less than 2, causes a kernel panic.