Interim fix for DMA "4K" problem.
The DMA triggers the rx callback when one of two things happen:
(a) The destination buffer is full.
(b) The inter-byte timeout expires (ageing timer).

When the imx.c operates at baud rates of 1200 or lower, and data is
sent at around 6-byte packets every 333ms, then the incoming bytes
are just spaced / spread wide enough to avoid the inter-byte timeout
from triggering. This results in the DMA engine not triggering the rx
callback until the 4K buffer is full.